WIP: First steps on implementing netCDF reader
.. note:: Tests have not yet been implemented

Starting to enable HazImp to directly read netcdf-format files. Under the
hood, HazImp uses `GDAL` to read raster data from a range of formats.
For most formats, this is fine, but to read NetCDF files, the filename
needs to explictly name the format, and also specify the variable name.

For example, for a NetCDF file named `hazard.nc` that contains three
variables 'temp', 'pressure' and 'wind' (not including dimension
variables), to load the 'wind' variable using the GDAL API requires the
following command:

  ds = gdal.Open('NETCDF:"hazard.nc":wind')

This new development adds additional configuration options under the
`local_wind` option:
- `file_format`: the extension of the file (e.g. 'nc' for netcdf)
- `variable`: the exact name of the variable to be loaded from the
netcdf file.

.. note:: No checking of the contents of the netcdf file is performed.
If the configured variable does not exist in the netcdf file, an error
will be raised.